+++ libbulletml/src/calc.h
@@ -7,7 +7,7 @@
 #include <string>
 #include <memory>
 
-DECLSPEC std::auto_ptr<Formula<double> > calc(const std::string& str);
+DECLSPEC std::auto_ptr<Formula<float> > calc(const std::string& str);
 
 #endif // CALC_H_
